nice pager controller

+++ App/Helpers/FFStyle.m
@@ -299,18 +299,18 @@ CGRect CGRectCopyWithOrigin(CGRect r, CGPoint origin)
 
 + (UIColor*)lighterColorForColor:(UIColor*)c
 {
-    float r, g, b, w, a;
+    CGFloat r, g, b, w, a;
     if ([c getRed:&r
              green:&g
               blue:&b
              alpha:&a]) {
-        return [UIColor colorWithRed:MIN(r + 0.2, 1.0)
-                               green:MIN(g + 0.2, 1.0)
-                                blue:MIN(b + 0.2, 1.0)
+        return [UIColor colorWithRed:MIN(r + 0.2f, 1.f)
+                               green:MIN(g + 0.2f, 1.f)
+                                blue:MIN(b + 0.2f, 1.f)
                                alpha:a];
     } else if ([c getWhite:&w
                       alpha:&a]) {
-        return [UIColor colorWithWhite:MIN(w + 0.2, 1.0)
+        return [UIColor colorWithWhite:MIN(w + 0.2f, 1.f)
                                  alpha:a];
     }
     return nil;
@@ -318,18 +318,18 @@ CGRect CGRectCopyWithOrigin(CGRect r, CGPoint origin)
 
 + (UIColor*)darkerColorForColor:(UIColor*)c
 {
-    float r, g, b, w, a;
+    CGFloat r, g, b, w, a;
     if ([c getRed:&r
              green:&g
               blue:&b
              alpha:&a]) {
-        return [UIColor colorWithRed:MAX(r - 0.2, 0.0)
-                               green:MAX(g - 0.2, 0.0)
-                                blue:MAX(b - 0.2, 0.0)
+        return [UIColor colorWithRed:MAX(r - 0.2f, 0.f)
+                               green:MAX(g - 0.2f, 0.f)
+                                blue:MAX(b - 0.2f, 0.f)
                                alpha:a];
     } else if ([c getWhite:&w
                       alpha:&a]) {
-        return [UIColor colorWithWhite:MAX(w - 0.2, 0.0)
+        return [UIColor colorWithWhite:MAX(w - 0.2f, 0.f)
                                  alpha:a];
     }
     return nil;
